requirements.txt
Flask
pandas
mysql-connector-python
flask-cors
python-dotenv
dash
plotly
statsmodels